moved progression logic to progression component
